feat: Drivers and users report form admin.

+++ voom-app/app/src/main/java/inc/visor/voom/app/admin/statistics/AdminStatisticsViewModel.java
@@ -32,8 +32,8 @@ public class AdminStatisticsViewModel extends ViewModel {
     public void loadReport(
             String from,
             String to,
-            Integer userId,
-            Integer driverId
+            Long userId,
+            Long driverId
     ) {
         loading.setValue(true);
